fixed filter buttons, other fixes
fixed the filter buttons by adding relative and absolute positions to the .box and circle icon as well as a left:0.

Cleaned up the accomodations section and added a popular section.

added a parent flexbox to hold both.

started working on a way to crop/scale images automatically. needs to be finished.