removed .img formating for images
removed instances of directly affecting img usages. This was not ideal for future additions of images.

added .crop class to crop and scale pictures for thumbnails for listings. The scaling needs to be finished.

a second image class needs to be added to adress the different structure in the popular category.