navbar changed, activities section added, footer added
changed the order and color of the .topnav contents. added anchors in html.
ISSUE
The links in .topnav need to be clicked twice to be directed to the appropriate anchor. Need to test further.

Added some simple objects to the activities section for testing. Need to add pictures and work on sizing.

Added and populated the footer. Links are non-functional. Footer text is centering, should be left justified.

Added footer to mediaquery and footer works as expected.